requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
httpx==0.27.2
pydantic==2.6.3
PyYAML==6.0.1
